add dependency injection library
