+++ .idea/gradle.xml
@@ -13,10 +13,15 @@
             <option value="$PROJECT_DIR$/core" />
             <option value="$PROJECT_DIR$/core/kotlin" />
             <option value="$PROJECT_DIR$/core/kotlin/coroutines" />
+            <option value="$PROJECT_DIR$/core/kotlin/coroutines/api" />
+            <option value="$PROJECT_DIR$/core/kotlin/coroutines/di" />
+            <option value="$PROJECT_DIR$/core/kotlin/coroutines/main" />
             <option value="$PROJECT_DIR$/core/kotlin/mapper" />
             <option value="$PROJECT_DIR$/core/network" />
+            <option value="$PROJECT_DIR$/core/network/api" />
             <option value="$PROJECT_DIR$/core/network/di" />
             <option value="$PROJECT_DIR$/core/network/main" />
+            <option value="$PROJECT_DIR$/core/network/model" />
             <option value="$PROJECT_DIR$/core/theme" />
             <option value="$PROJECT_DIR$/data" />
             <option value="$PROJECT_DIR$/data/base-feature" />
