+++ app/src/main/java/com/kacek/sandbox/MainActivity.kt
@@ -8,9 +8,11 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Scaffold
 import androidx.compose.ui.Modifier
-import com.kacek.ui.basefeature.BaseScreen
 import com.kacek.theme.AppTheme
+import com.kacek.ui.basefeature.BaseScreen
+import dagger.hilt.android.AndroidEntryPoint
 
+@AndroidEntryPoint
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
