+++ app/src/main/java/com/kacek/sandbox/MainApplication.kt
@@ -1,0 +1,7 @@
+package com.kacek.sandbox
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class MainApplication : Application()
