+++ core/kotlin/coroutines/api/src/main/java/com/kacek/core/kotlin/coroutines/api/providers/DispatcherProvider.kt
@@ -1,0 +1,10 @@
+package com.kacek.core.kotlin.coroutines.api.providers
+
+import kotlin.coroutines.CoroutineContext
+
+interface DispatcherProvider {
+    val default: CoroutineContext
+    val main: CoroutineContext
+    val io: CoroutineContext
+    val unconfined: CoroutineContext
+}
