--- core/kotlin/coroutines/build.gradle.kts
@@ -1,7 +1,0 @@
-plugins {
-    id("kotlin")
-}
-
-dependencies {
-    implementation(libs.coroutines)
-}
