+++ core/kotlin/coroutines/di/build.gradle.kts
@@ -5,8 +5,7 @@ plugins {
 
 dependencies {
     implementation(project(":core:kotlin:coroutines:api"))
-    implementation(project(":core:network:api"))
-    implementation(project(":core:network:main"))
+    implementation(project(":core:kotlin:coroutines:main"))
 
     implementation(libs.dagger.hilt.core)
     kapt(libs.dagger.hilt.compiler)
