+++ core/kotlin/coroutines/di/src/main/java/com/kacek/core/kotlin/coroutines/di/CoroutinesModule.kt
@@ -1,0 +1,18 @@
+package com.kacek.core.kotlin.coroutines.di
+
+import com.kacek.core.kotlin.coroutines.api.providers.DispatcherProvider
+import com.kacek.core.kotlin.coroutines.providers.CoroutineDispatcherProvider
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+
+@Module
+@InstallIn(SingletonComponent::class)
+object CoroutinesModule {
+
+    @Provides
+    fun provideDispatcherProvider(): DispatcherProvider {
+        return CoroutineDispatcherProvider()
+    }
+}
