+++ core/kotlin/coroutines/main/build.gradle.kts
@@ -1,0 +1,9 @@
+plugins {
+    id("kotlin")
+}
+
+dependencies {
+    implementation(project(":core:kotlin:coroutines:api"))
+
+    implementation(libs.coroutines)
+}
