--- core/kotlin/coroutines/src/main/java/com/kacek/core/kotlin/coroutines/providers/DispatcherProvider.kt
@@ -1,10 +1,0 @@
-package com.kacek.core.kotlin.coroutines.providers
-
-import kotlin.coroutines.CoroutineContext
-
-interface DispatcherProvider {
-    val default: CoroutineContext
-    val main: CoroutineContext
-    val io: CoroutineContext
-    val unconfined: CoroutineContext
-}
