+++ core/network/api/build.gradle.kts
@@ -1,0 +1,7 @@
+plugins {
+    id("kotlin")
+}
+
+dependencies {
+    implementation(project(":core:network:model"))
+}
