--- core/network/di/src/main/java/com/kacek/network/di/MyClass.kt
@@ -1,4 +1,0 @@
-package com.kacek.network.di
-
-class MyClass {
-}
