--- core/network/main/src/main/java/com/kacek/network/DummyNetwork.kt
@@ -1,3 +1,0 @@
-package com.kacek.network
-
-class DummyNetwork
