--- core/network/main/src/main/java/com/kacek/network/mapper/KtorHttpResponseToResponseMapper.kt
@@ -1,22 +1,0 @@
-package com.kacek.network.mapper
-
-import com.kacek.network.model.response.Response
-import com.kacek.network.model.response.ResponseTypeInfo
-import io.ktor.client.call.body
-import io.ktor.client.statement.HttpResponse
-import io.ktor.util.reflect.TypeInfo
-
-internal class KtorHttpResponseToResponseMapper {
-
-    suspend fun <T> map(
-        httpResponse: HttpResponse,
-        responseType: ResponseTypeInfo,
-    ): Response<T> {
-        return Response(
-            statusCode = httpResponse.status.value,
-            content = httpResponse.body(responseType.toKtorTypeInfo()),
-        )
-    }
-
-    private fun ResponseTypeInfo.toKtorTypeInfo() = TypeInfo(type, reifiedType, kotlinType)
-}
