--- core/network/main/src/main/java/com/kacek/network/model/response/Response.kt
@@ -1,6 +1,0 @@
-package com.kacek.network.model.response
-
-data class Response<T>(
-    val statusCode: Int,
-    val content: T,
-)
