+++ core/network/model/build.gradle.kts
@@ -1,0 +1,3 @@
+plugins {
+    id("kotlin")
+}
