+++ core/network/model/src/main/java/com/kacek/core/network/model/response/Response.kt
@@ -1,0 +1,6 @@
+package com.kacek.core.network.model.response
+
+data class Response<T>(
+    val statusCode: Int,
+    val content: T,
+)
