+++ settings.gradle.kts
@@ -21,10 +21,14 @@ dependencyResolutionManagement {
 
 rootProject.name = "Android-kacek-sandbox"
 include(":app")
-include(":core:kotlin:coroutines")
+include(":core:kotlin:coroutines:main")
+include(":core:kotlin:coroutines:api")
+include(":core:kotlin:coroutines:di")
 include(":core:kotlin:mapper")
+include(":core:network:api")
 include(":core:network:di")
 include(":core:network:main")
+include(":core:network:model")
 include(":core:theme")
 include(":data:base-feature")
 include(":domain:base-feature")
